Move the RenderCommand into the shared draw module
